+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/model/dto/UserDtos.kt
@@ -79,6 +79,7 @@ internal data class DownstreamUserDto(
     val channel_mutes: List<DownstreamChannelMuteDto>?,
     val blocked_user_ids: List<String>?,
     val avg_response_time: Long?,
+    val push_preferences: DownstreamPushPreferenceDto?,
     val extraData: Map<String, Any>,
 ) : ExtraDataDto